+++ services/web-app/components/catalog-search/catalog-search.js
@@ -5,20 +5,25 @@
  * LICENSE file in the root directory of this source tree.
  */
 import { Column, Grid, Layer, Search, Theme } from '@carbon/react'
-import { useState } from 'react'
 
 import CatalogMultiselectFilter from '@/components/catalog-multislect-filter'
-import { mediaQueries, useMatchMedia } from '@/utils/media-query'
+import { mediaQueries, useMatchMedia } from '@/utils/use-match-media'
 
 import styles from './catalog-search.module.scss'
 
-const CatalogSearch = ({ onSelect, onSearch }) => {
-  const [query, setQuery] = useState('')
+const CatalogSearch = ({ search = '', onSearch, onSelect }) => {
   const isMd = useMatchMedia(mediaQueries.md)
 
+  const handleOnBlur = (event) => {
+    onSearch(event.target.value, true)
+  }
+
   const handleOnChange = (event) => {
-    setQuery(event.target.value)
-    onSearch(event.target.value)
+    onSearch(event.target.value, false)
+  }
+
+  const handleOnClear = () => {
+    onSearch('', true)
   }
 
   return (
@@ -30,8 +35,10 @@ const CatalogSearch = ({ onSelect, onSearch }) => {
               id="catalog-search"
               labelText="Search component index by name, keyword, or domain"
               placeholder="Component name, keyword, domain"
-              value={query}
+              value={search}
+              onBlur={handleOnBlur}
               onChange={handleOnChange}
+              onClear={handleOnClear}
               size="lg"
             />
           </Column>