+++ services/web-app/pages/assets/components.js
@@ -11,7 +11,6 @@ import Catalog from '@/components/catalog'
 import { assetsNavData } from '@/data/nav-data'
 import { LayoutContext } from '@/layouts/layout'
 import { getAllLibraries } from '@/lib/github'
-import styles from '@/pages/pages.module.scss'
 
 const Components = ({ librariesData }) => {
   const { setNavData } = useContext(LayoutContext)
@@ -28,7 +27,6 @@ const Components = ({ librariesData }) => {
     <>
       <NextSeo {...seo} />
       <Catalog data={librariesData} type="component" />
-      <pre className={styles.data}>{JSON.stringify(librariesData, null, 2)}</pre>
     </>
   )
 }