+++ services/web-app/utils/use-event-listener.js
@@ -1,0 +1,35 @@
+/*
+ * Copyright IBM Corp. 2021, 2021
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import { useEffect, useRef } from 'react'
+
+export const useEventListener = (eventName, handler, element) => {
+  if (!element && typeof window !== 'undefined') {
+    element = window
+  }
+
+  const savedHandler = useRef()
+
+  useEffect(() => {
+    savedHandler.current = handler
+  }, [handler])
+
+  useEffect(() => {
+    const isSupported = element && element.addEventListener
+    if (!isSupported) return
+
+    const eventListener = (event) => savedHandler.current(event)
+
+    element.addEventListener(eventName, eventListener)
+
+    return () => {
+      element.removeEventListener(eventName, eventListener)
+    }
+  }, [eventName, element])
+}
+
+export default useEventListener